week_2_data_ingestion/homework/requirements.txt
apache-airflow-providers-google
pyarrow
pandas 
sqlalchemy
psycopg2-binary
oauth2client